changes into HomeFragment from NewsFragment (now blank)
