+++ app/src/main/java/com/melvin/ongandroid/view/HomeFragment.kt
@@ -9,20 +9,25 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.viewModels
+import androidx.lifecycle.Observer
+import androidx.recyclerview.widget.LinearLayoutManager
 import com.melvin.ongandroid.R
 import com.melvin.ongandroid.businesslogic.data.DataSource
 import com.melvin.ongandroid.businesslogic.domain.OnRegister
+import com.melvin.ongandroid.businesslogic.vo.Resource
 import com.melvin.ongandroid.databinding.FragmentHomeBinding
 import com.melvin.ongandroid.databinding.FragmentLoginBinding
+import com.melvin.ongandroid.model.New
 import com.melvin.ongandroid.model.repository.RepoImpl
+import com.melvin.ongandroid.viewmodel.HomeViewModel
 import com.melvin.ongandroid.viewmodel.UserViewModel
 import com.melvin.ongandroid.viewmodel.VMFactory
 
 
-class HomeFragment : Fragment() {
+class HomeFragment : Fragment(),NewsAdapter.OnNewClickListener {
     private var _binding: FragmentHomeBinding? = null
     private val binding get() = _binding!!
-
+    private val viewModel by viewModels<HomeViewModel>(){ VMFactory(RepoImpl(DataSource())) }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -37,13 +42,40 @@ class HomeFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         _binding = FragmentHomeBinding.inflate(inflater, container, false)
-      //  return inflater.inflate(R.layout.fragment_home, container, false)
+
+        setUpRecyclerView()
+
+        viewModel.fetchNewsList.observe(viewLifecycleOwner, Observer { result ->
+            when(result){
+                is Resource.Loading->{
+                    binding.prBar.visibility=View.VISIBLE
+                    binding.prError.visibility=View.GONE
+                }
+                is Resource.Success->{
+                    binding.prBar.visibility=View.GONE
+                    binding.prError.visibility=View.GONE
+                    binding.rvNews.adapter=NewsAdapter(requireContext(), result.data, this)
+                }
+                is Resource.Failure->{
+                    binding.prBar.visibility=View.GONE
+                    binding.prError.visibility=View.VISIBLE
+                    Toast.makeText(requireContext(),R.string.login_dg_without_internet,Toast.LENGTH_LONG).show()
+                }
+
+            }
+        })
 
         return binding.root
 
     }
 
+    private fun setUpRecyclerView() {
+        val appContext = requireContext().applicationContext
+        val recyclerView = binding.rvNews
+        recyclerView.layoutManager = LinearLayoutManager(appContext, LinearLayoutManager.HORIZONTAL, false)
+    }
 
-
-
+    override fun onNewClick(new: New) {
+        Toast.makeText(requireContext(),R.string.news_coming,Toast.LENGTH_LONG).show()
+    }
 }
