+++ app/src/main/java/com/melvin/ongandroid/view/NewsFragment.kt
@@ -12,18 +12,17 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.melvin.ongandroid.R
 import com.melvin.ongandroid.businesslogic.data.DataSource
 import com.melvin.ongandroid.businesslogic.vo.Resource
-import com.melvin.ongandroid.databinding.FragmentHomeBinding
 import com.melvin.ongandroid.databinding.FragmentNewsBinding
 import com.melvin.ongandroid.model.New
 import com.melvin.ongandroid.model.repository.RepoImpl
-import com.melvin.ongandroid.viewmodel.NewsViewModel
+import com.melvin.ongandroid.viewmodel.HomeViewModel
 import com.melvin.ongandroid.viewmodel.VMFactory
 
 
-class NewsFragment : Fragment(), NewsAdapter.OnNewClickListener {
+class NewsFragment : Fragment() {
     private var _binding: FragmentNewsBinding? = null
     private val binding get() = _binding!!
-    private val viewModel by viewModels<NewsViewModel>(){ VMFactory(RepoImpl(DataSource())) }
+
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -34,40 +33,10 @@ class NewsFragment : Fragment(), NewsAdapter.OnNewClickListener {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentNewsBinding.inflate(inflater, container, false)
 
-        setUpRecyclerView()
-
-        viewModel.fetchNewsList.observe(viewLifecycleOwner, Observer { result ->
-            when(result){
-                is Resource.Loading->{
-                    binding.prBar.visibility=View.VISIBLE
-                    binding.prError.visibility=View.GONE
-                }
-                is Resource.Success->{
-                    binding.prBar.visibility=View.GONE
-                    binding.prError.visibility=View.GONE
-                    binding.rvNews.adapter=NewsAdapter(requireContext(), result.data, this)
-                }
-                is Resource.Failure->{
-                    binding.prBar.visibility=View.GONE
-                    binding.prError.visibility=View.VISIBLE
-                    Toast.makeText(requireContext(),R.string.login_dg_without_internet,Toast.LENGTH_LONG).show()
-                }
-
-            }
-        })
 
         return binding.root
     }
 
-    private fun setUpRecyclerView() {
-        val appContext = requireContext().applicationContext
-        val recyclerView = binding.rvNews
-        recyclerView.layoutManager = LinearLayoutManager(appContext)
-    }
 
-    override fun onNewClick(new: New) {
-        Toast.makeText(requireContext(),R.string.news_coming,Toast.LENGTH_LONG).show()
-    }
 }
