+++ app/src/main/res/layout/fragment_news.xml
@@ -52,6 +52,8 @@
         android:id="@+id/rvNews"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:orientation="horizontal"
+        tools:listitem="@layout/news_row"
 
         >
